termproxy: fix parsing free-form listen-port argument
While commit d8e9c1a ("termproxy: fix the command line help text")
corrected the usage text to match the behavior and was a good stop-gap
it simply is to easy to make this work as most CLI tools do and most
user expect.

Switch to parsing all options first and only then get the free-form
listen-port(-or-fd) argument.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>