termproxy: avoid gluing words together in CLI option names
Separate auth-port and auth-socket with a minus to make them easier to
read. Fallback to the old style for `authport` which is pre-existing
and used by, e.g. PBS. As the second option, now named `auth-socket`,
got only introduced very recently and was not shipped in any package,
we can ignore backwards compat for it.

Update and slightly rework the usage output too.

Note that even though this is an internal tool the cost doing so is
small and it can only help, e.g. on using this manually and/or during
debugging.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>