xtermjs: upgrade to 5.5.0
and upgrade fit-addon to 0.9.0 and webgl-addon to 0.17.0 to be
compatible.

We have to slightly adapt the Makefile to download the correct files, as
xtermjs now uses a different package naming:

xterm -> @xterm/xterm
xterm-addon-* -> @xterm/addon-*

which also changes the extracted file names from xterm-addon-*.js to
addon-*.js so adapt them in the index templates.

This also means we have to change the lintian override paths for those.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>