deregister stream from temporary poll handle
In mio 1 it is required that sources get deregistered from a registry
before registering them somewhere else, otherwise it errors that the
handle is already registered somewhere, despite the Poll instance
having been dropped...

This is highly inconvenient and it does not provide guards for this.
Fortunately we only have 1 such case.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>