Avoid allocating in `pre_exec` closure
`Error::new` allocates memory (see [148971]). This is bad in
multi-threaded programs, which pve-xtermjs AFAIK is. If the fork occurs
while the allocator lock is held by another thread, deadlocks can occur,
since there's no one left in the new process to unlock the mutex.
I do not believe this is UB, and modern libc offer protections against
this issue, but this isn't POSIX-compliant and should preferably be
avoided.

`nix` provides a non-allocating `impl From<Errno> for std::io::Error`,
which can be used instead. This ensures that the correct error code is
forwarded to the parent process, instead of the default `-EINVAL`.

[rust#148971]: https://github.com/rust-lang/rust/pull/148971

Signed-off-by: Yuki Sireneva <yuki.utk8g@gmail.com>
Link: https://lore.proxmox.com/20251205124802.334198-1-yuki.utk8g@gmail.com